webapp/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
django-debug-toolbar==4.2.0
Pillow

